card de lista nominal com nova estrutura
